Check for files in sdist but not in VCS again (#27)
This check requires that we build an sdist directly in the original
source tree, without making a clean copy.  But we need a clean copy in
order to check MANIFEST.in when setuptools VCS logic (or legacy
SOURCES.txt parsing) kicks in!  To satisfy both requirements we end up
having to build the sdist twice: once in a clean tree, and once in the
original, dirty tree.

The

  missing_from_VCS = set(sdist_files + clean_sdist_files) - set(source_files)

check could probably have been written as

  missing_from_VCS = set(sdist_files) - set(source_files)

because I'm 99.9% sure sdist_files is a superset of clean_sdist_files.
But just in case I'm wrong I thought it's better to be sure.