Allow for reading configuration from a pyproject.toml file
PEP 518 has introduced of pyproject.toml and it has been adopted by
other tools as the standard location for tooling configuration. This
allows for the check-manifest configuration to also be loaded from the
pyproject.toml file. Note this requires ``toml`` to be installed,
which is added as an optional extra.