Always require toml
It's a small package with no external depencencies.  It'll be more
convenient for users who have pyproject.toml.  And this way I can go
back to 100% test coverage without figuring out how to test both the
"toml is missing during import time" and "toml is present during import
time" branches in the code.