tox 4 requires passenv to be newline (or comma) separated
tox 3 requires passenv to be space (or newline) separated.

Using newlines works on both.

I'm not sure the passenv is still needed here.  I'm pretty sure I had to
add it back when I was running my own Jenkins builds on Windows.  Oh
well, the last Appveyor build happened 6 days ago when tox 4 was not out
yet, so won't know if it would've broken or not.

(Ok, LANG/LC_CTYPE/LC_ALL was for Python versions that used the locale
instead of defaulting to UTF-8, so I could create files with non-ASCII
filenames in the filesystem.  I think the currently supported Python
versions don't need those any more.  MSYSTEM was for Git for Windows to,
uh, do something like support symlinks?  Or print filenames with forward
slashes?  That one was for my Windows Jenkins builds.)