Clarify docs for --no-build-isolation
Stop assuming that setuptools/wheel will always be preinstalled and
don't need to be explicitly listed as build requirements; this is not
the case in Python 3.12.

Stop using ArgumentDefaultsHelpFormatter, it makes the help message for
--no-build-isolation extra confusing.

Clarify that 'patterns' in --ignore et al. mean glob patterns rather
than regexes or something.

Use an exact version instead of `... # pick a valid tag / revision` in
the README, I trust my `make check-readme` (part of `make release`) to
remind me to update it.

See #168.