Make 'make check-readme' check both rev: lines
If either one of them differs from the expected value, print an error
and exit.

(Previously the Makefile rule would be satisfied even if only one of the
two rev: lines were correct.)

I'm sure I'm going to regret going all-in on this shell one-liner at
some point in the future when I have to debug it.